+++ src/cli.rs
@@ -17,10 +17,11 @@ use pcli2::{
     actions::{
         assets::{
             create_asset, create_asset_batch, create_asset_metadata_batch, delete_asset,
-            delete_asset_metadata, download_asset, geometric_match_asset, geometric_match_folder,
-            list_assets, metadata_inference, part_match_asset, part_match_folder, print_asset,
-            print_asset_dependencies, print_asset_metadata, print_folder_dependencies, reprocess_asset,
-            text_match, update_asset_metadata, visual_match_asset, visual_match_folder,
+            delete_asset_metadata, download_asset, download_asset_thumbnail, geometric_match_asset,
+            geometric_match_folder, list_assets, metadata_inference, part_match_asset,
+            part_match_folder, print_asset, print_asset_dependencies, print_asset_metadata,
+            print_folder_dependencies, reprocess_asset, text_match, update_asset_metadata,
+            visual_match_asset, visual_match_folder,
         },
         folders::{
             create_folder, delete_folder, download_folder, list_folders, move_folder,
@@ -38,11 +39,11 @@ use pcli2::{
             COMMAND_CREATE, COMMAND_CREATE_BATCH, COMMAND_CURRENT, COMMAND_DELETE,
             COMMAND_DEPENDENCIES, COMMAND_DOWNLOAD, COMMAND_EXPORT, COMMAND_FOLDER, COMMAND_GET,
             COMMAND_IMPORT, COMMAND_INFERENCE, COMMAND_LIST, COMMAND_LOGIN, COMMAND_LOGOUT,
-            COMMAND_MATCH, COMMAND_METADATA, COMMAND_PART_MATCH, COMMAND_REPROCESS, COMMAND_STATE, COMMAND_TENANT,
-            COMMAND_TEXT_MATCH, COMMAND_UPLOAD, COMMAND_USE, COMMAND_VISUAL_MATCH,
-            PARAMETER_API_URL, PARAMETER_AUTH_URL, PARAMETER_CLIENT_ID, PARAMETER_CLIENT_SECRET,
-            PARAMETER_FILE, PARAMETER_FORMAT, PARAMETER_HEADERS, PARAMETER_OUTPUT,
-            PARAMETER_PRETTY, PARAMETER_UI_URL,
+            COMMAND_MATCH, COMMAND_METADATA, COMMAND_PART_MATCH, COMMAND_REPROCESS, COMMAND_STATE,
+            COMMAND_TENANT, COMMAND_TEXT_MATCH, COMMAND_THUMBNAIL, COMMAND_UPLOAD, COMMAND_USE,
+            COMMAND_VISUAL_MATCH, PARAMETER_API_URL, PARAMETER_AUTH_URL, PARAMETER_CLIENT_ID,
+            PARAMETER_CLIENT_SECRET, PARAMETER_FILE, PARAMETER_FORMAT, PARAMETER_HEADERS,
+            PARAMETER_OUTPUT, PARAMETER_PRETTY, PARAMETER_UI_URL,
         },
     },
     format::{Formattable, FormattingError, OutputFormat, OutputFormatOptions},
@@ -390,6 +391,11 @@ pub async fn execute_command() -> Result<(), CliError> {
                             sub_matches,
                         ))),
                     }
+                }
+                Some((COMMAND_THUMBNAIL, sub_matches)) => {
+                    trace!("Command: {} {}", COMMAND_ASSET, COMMAND_THUMBNAIL);
+                    download_asset_thumbnail(sub_matches).await?;
+                    Ok(())
                 }
                 Some((COMMAND_REPROCESS, sub_matches)) => {
                     trace!("Command: {} {}", COMMAND_ASSET, COMMAND_REPROCESS);